leave_system-main/leave_system/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
SpiffWorkflow==2.0.1
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1
celery==5.3.4
redis==5.0.1
